requirements.txt
opencv-python==4.10.0.84
ultralytics==8.3.5
mediapipe==0.10.14
google-generativeai==0.8.3
PyQt5==5.15.11
pyttsx3==2.91
PyPDF2==3.0.1
reportlab==4.2.2
scipy==1.14.1
numpy>=1.26.4
bcrypt==4.2.0
twilio==9.3.0